+++ packages/xrpl/HISTORY.md
@@ -13,7 +13,10 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 * Fixed the location of `signer_lists` in the `account_info` response so that it matches rippled
 * Guard check for signing algorithm used in `Wallet.generate()`
 * Null and undefined values in transactions are now treated as though the field was not passed in.
-* Improved the type definition of the return value of `submitAndWait()`.
+* Improved the type definition of the return value of `submitAndWait()`
+
+### Changed
+* Added sidechain devnet support to faucet generation
 
 ### Removed
 * RPCs and utils related to the old sidechain design